channel: migrate tests for awaiting_funding_depth transitions
