+++ packages/lightning/__tests__/channels/TransitionFactory.spec.ts
@@ -12,8 +12,10 @@ import { Result } from "../../lib/Result";
 import { TransitionFactory } from "../../lib";
 import {
     createFakeAcceptChannel,
+    createFakeBlock,
     createFakeChannel,
     createFakeChannelLogicFacade,
+    createFakeChannelReady,
     createFakeFundingCreatedMessage,
     createFakeFundingSignedMessage,
     createFakeFundingTx,
@@ -193,6 +195,158 @@ describe(TransitionFactory.name, () => {
 
                 // assert
                 expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingFundingDepth);
+            });
+        });
+    });
+
+    describe(TransitionFactory.prototype.createOnBlockConnected.name, () => {
+        let logger: ILogger;
+        let logic: sinon.SinonStubbedInstance<IChannelLogic>;
+        let sut: TransitionFactory;
+        let channel: Channel;
+
+        beforeEach(() => {
+            logger = createFakeLogger();
+            logic = createFakeChannelLogicFacade();
+            sut = new TransitionFactory(logger, logic);
+            channel = createFakeChannel()
+                .attachAcceptChannel(createFakeAcceptChannel())
+                .attachFundingTx(createFakeFundingTx())
+                .attachFundingSigned(createFakeFundingSignedMessage());
+        });
+
+        it("not confirmed => stays", async () => {
+            // arrange
+            const block = createFakeBlock();
+            const event = new ChannelEvent(ChannelEventType.BlockConnected);
+            event.block = block;
+
+            // act
+            const result = await sut.createOnBlockConnected()(channel, event);
+
+            // assert
+            expect(channel.fundingConfirmedHeight).to.equal(undefined);
+            expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingFundingDepth);
+        });
+
+        it("confirmed => attaches height, stays", async () => {
+            // arrange
+            const fundingTx = createFakeFundingTx();
+            const block = createFakeBlock(500_000, fundingTx);
+            const event = new ChannelEvent(ChannelEventType.BlockConnected);
+            event.block = block;
+
+            // act
+            const result = await sut.createOnBlockConnected()(channel, event);
+
+            // assert
+            expect(channel.fundingConfirmedHeight).to.equal(500_000);
+            expect(channel.readyHeight).to.equal(500_005);
+            expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingFundingDepth);
+        });
+
+        it("confirmed + height => stays", async () => {
+            // arrange
+            channel.markConfirmed(500_000);
+            const block = createFakeBlock(500_001);
+            const event = new ChannelEvent(ChannelEventType.BlockConnected);
+            event.block = block;
+
+            // act
+            const result = await sut.createOnBlockConnected()(channel, event);
+
+            // assert
+            expect(channel.fundingConfirmedHeight).to.equal(500_000);
+            expect(channel.readyHeight).to.equal(500_005);
+            expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingFundingDepth);
+        });
+
+        it("meets depth, no channel_ready => sends channel_ready + transitions to awaiting_channel_ready", async () => {
+            // arrange
+            channel.markConfirmed(500_000);
+            const block = createFakeBlock(500_005);
+            const event = new ChannelEvent(ChannelEventType.BlockConnected);
+            event.block = block;
+
+            // act
+            const result = await sut.createOnBlockConnected()(channel, event);
+
+            // assert
+            expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingChannelReady);
+            expect(logic.sendMessage.called).to.equal(true);
+        });
+
+        it("meets depth + has channel_ready => sends channel_ready + transitions to normal", async () => {
+            // arrange
+            channel.markConfirmed(500_000);
+            channel.attachChannelReady(createFakeChannelReady());
+
+            const block = createFakeBlock(500_006);
+            const event = new ChannelEvent(ChannelEventType.BlockConnected);
+            event.block = block;
+
+            // act
+            const result = await sut.createOnBlockConnected()(channel, event);
+
+            // assert
+            expect(result).to.equal(ChannelStateId.Channel_Normal);
+            expect(logic.sendMessage.called).to.equal(true); // arrange
+        });
+    });
+
+    describe(TransitionFactory.prototype.createOnAcceptChannelMessageTransition.name, () => {
+        let logger: ILogger;
+        let logic: sinon.SinonStubbedInstance<IChannelLogic>;
+        let sut: TransitionFactory;
+        let channel: Channel;
+
+        beforeEach(() => {
+            logger = createFakeLogger();
+            logic = createFakeChannelLogicFacade();
+            sut = new TransitionFactory(logger, logic);
+            channel = createFakeChannel()
+                .attachAcceptChannel(createFakeAcceptChannel())
+                .attachFundingTx(createFakeFundingTx())
+                .attachFundingSigned(createFakeFundingSignedMessage());
+        });
+
+        describe("in awaiting_funding_dpeth", () => {
+            it("transitions Failed when invalid", async () => {
+                // arrange
+                const msg = createFakeChannelReady();
+                const event = new ChannelEvent(ChannelEventType.ChannelReadyMessage);
+                event.message = msg;
+                logic.validateChannelReadyMessage.returns(false);
+
+                // act
+                const result = await sut.createOnChannelReadyTransition(
+                    ChannelStateId.Channel_Opening_AwaitingFundingDepth,
+                )(channel, event);
+
+                // assert
+                expect(result).to.equal(ChannelStateId.Channel_Failing);
+            });
+
+            it("attaches next_per_commitment_point + stays", async () => {
+                // arrange
+                const msg = createFakeChannelReady();
+                const event = new ChannelEvent(ChannelEventType.ChannelReadyMessage);
+                event.message = msg;
+                logic.validateChannelReadyMessage.returns(true);
+
+                // act
+                const result = await sut.createOnChannelReadyTransition(
+                    ChannelStateId.Channel_Opening_AwaitingFundingDepth,
+                )(channel, event);
+                expect(result).to.equal(ChannelStateId.Channel_Opening_AwaitingFundingDepth);
+                expect(channel.theirSide.commitmentNumber.value).to.equal(0n);
+                expect(channel.theirSide.commitmentPoint.toHex()).to.equal(
+                    "0288a618cb6027c3218a37cbe9e882379f17d87d03f6e99d0b60292478d2aded06",
+                );
+                expect(channel.theirSide.nextCommitmentNumber.value).to.equal(1n);
+                expect(channel.theirSide.nextCommitmentPoint.toHex()).to.equal(
+                    "032405cbd0f41225d5f203fe4adac8401321a9e05767c5f8af97d51d2e81fbb206",
+                );
             });
         });
     });
